app/src/main/java/com/larissa/desafio_android_larissa_carvalho/network/ClientAPI.java
package com.larissa.desafio_android_larissa_carvalho.network;

import com.larissa.desafio_android_larissa_carvalho.utils.Constants;

import retrofit2.Retrofit;
import retrofit2.converter.gson.GsonConverterFactory;

public class ClientAPI {

    private static Retrofit retrofit;

    public static Retrofit getRetrofit() {
        if (retrofit == null){
            retrofit = new Retrofit.Builder()
                    .baseUrl(Constants.BASE_URL)
                    .addConverterFactory(GsonConverterFactory.create())
                    .build();
        }

        return retrofit;
    }
}
